WIP: Added concurrency between input handling and ui
There is still a ton of issues and considerations for supporting a
concurrent architecture on this project.

New to the project is using `Fiber`s for concurrency work handling.
There are 2 fibers introduced call `Rock::Foreman` and `Rock::Screen`.
Foreman handles and parses the incoming data for user input. Screen
handles the drawing of `Rock` to an output; at this time `Rock` is only
interfacing within an xterm-like virtual terminal.

There are other file additions and changes but nothing of note atm.

The primary goal is to learn and implement the concurrency patterns
within Crystal to build a quick application while achieving a cohesive
codebase. This is a decent starting point, but there are many other
challenges ahead for other components.

There is an issue regarding the blocking aspect of `Channel#receive`
that is difficult to workaround. When a fiber calls any
`Channel#receive`, it will be blocked until the corresponding
`Channel#send` has send the data. For certain fibers, we might want to
keep executing and not wait for communication from a `Channel`.

For example, the UI fiber only resumes once the Foreman fiber has
processed any input from the user and called `Channel#send`.

In certain contexts, it maybe preferable to block execution of some
fibers from wasting cycles when there hasn't been any changes, such as
not having the UI fiber to not draw when there's no change.