Fix: Stopped paging output to the terminal screen
Incorrectly used ESC [ 2J sequence when clearing what the user sees
in the terminal before writing the content.

The only time we want to use `ESC [ 2J` is on initial render to erase
(it's more like 'page up' or `CTRL-l` in the shell) the shell content.

However, there may be an alternative to avoid this behaviour by using
xterm's alternative buffer feature, but needs more research to see if
this is what nvim uses for displaying.