Refactor: Removing Quarry and Radio alias
Removing Quarry for the time being, since the main issue with holding a
list of channels had caveats with Crystal's type system for Channel
types, since the list will use a Union type for the value, it made
necessary to write explicit type code when fetching a channel.

This might be fine since the compiler catches mismatch types but right
now `Rock` extensively uses Crystal's namespacing features and writing
the type in certain scopes becomes very tedious and verbose.

Will revisit this concept in the future when needing to scale management
of channels.