Feat: #4 and #5 initial modules for window and content
Setting up the base objects for the layers that will manage the content
for user interactions.

Right now the `Screen` module will handle the window management.

The `Bench` class is the equivalent to a pane in tmux. Its purpose right
now is to act as a container for the content, in the future multiple
benches will be able to split the view for displaying multiple files as
an example.

The `Face` class is the container for content. It holds a `IO::Memory`
buffer for storing the user's input when the editor is in insert mode.

The `Screen` module for the time being will be the window manager. As
well it will setup the keymaps to enable adding text to the right
location for the active bench.