Feat: #5 Refactored `Rock::Document` to use singular buffers
Entries in a piece table in general should only reference index locations to a
particular content buffer.

Right now there is a single buffer for the original content (needs to be
implemented) and the edit content. The `Rock::Entry` class will specify
the buffer where its content is located, the start index in the buffer,
and the number of bytes. `Document` holds a list of `Entry` items that
are in order on insert or updates.

Before these changes, `Document` was more logically similar to a gap
buffer. Each entry referenced the start position that would relate more
to the position it should be in the final content sequence. But when
trying to implement a window (or range) for getting a selection of the
content, by specifying the range of new lines, it was becoming very
difficult to calculate the index positions for getting all the bytes in
all cases.

It's unclear if the same issue or challenge will happen with the proper
logic for a piece table data structure.