Feat: #5 `to_slice` method for a range of lines added to `Document`
Since there is will be a limited number of lines that a buffer would
be able to display, the `Document` object can now return a specified range
of content to avoid having to do retrieving everything.

Now this is only theoretical because I don't have a better idea of
efficiently keeping track of content changes and rendering to screen.

I'm hoping this stage of the pipeline (as well, the pipeline in general)
isn't too slow and not cause significant latency or flickering on basic
usage.